test: move materialized view alter basics to internal package
