+++ pkg/executor/test/ddl/materialized_view_basic_test.go
@@ -22,6 +22,7 @@ import (
 	"testing"
 
 	"github.com/pingcap/tidb/pkg/ddl"
+	"github.com/pingcap/tidb/pkg/errno"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/meta/model"
 	"github.com/pingcap/tidb/pkg/parser/ast"
@@ -429,4 +430,397 @@ func TestDropMaterializedViewIfExists(t *testing.T) {
 	tk.MustQuery("show warnings").Check(testkit.Rows("Note 1051 Unknown table 'test.t_no_mlog_drop_if_exists'"))
 	tk.MustExec("drop materialized view log if exists on t_drop_if_exists")
 	tk.MustExec("drop materialized view log if exists on t_drop_if_exists")
+}
+
+func TestAlterMaterializedViewRefreshExprTypeValidation(t *testing.T) {
+	store := testkit.CreateMockStore(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t (a int not null, b int not null)")
+	tk.MustExec("create materialized view log on t (a, b)")
+	tk.MustExec("create materialized view mv (a, s, cnt) refresh fast next date_add(now(), interval 1 hour) as select a, sum(b), count(1) from t group by a")
+
+	err := tk.ExecToErr("alter materialized view mv refresh next 300")
+	require.ErrorContains(t, err, "REFRESH NEXT expression must return DATETIME/TIMESTAMP")
+
+	err = tk.ExecToErr("alter materialized view mv refresh start with 1 next date_add(now(), interval 1 hour)")
+	require.ErrorContains(t, err, "REFRESH START WITH expression must return DATETIME/TIMESTAMP")
+
+	tk.MustExec("alter materialized view mv refresh start with now() next date_add(now(), interval 1 hour)")
+}
+
+func TestAlterMaterializedViewMetadataDDL(t *testing.T) {
+	store, dom := testkit.CreateMockStoreAndDomain(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t (a int not null, b int not null)")
+	tk.MustExec("insert into t values (1, 10), (1, 5), (2, 7)")
+	tk.MustExec("create materialized view log on t (a, b) purge next date_add(now(), interval 1 hour)")
+	tk.MustExec("create materialized view mv (a, s, cnt) refresh fast next date_add(now(), interval 2 hour) as select a, sum(b), count(1) from t group by a")
+
+	getMView := func() (*model.TableInfo, *model.MaterializedViewInfo) {
+		is := dom.InfoSchema()
+		tbl, err := is.TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("mv"))
+		require.NoError(t, err)
+		require.NotNil(t, tbl.Meta().MaterializedView)
+		return tbl.Meta(), tbl.Meta().MaterializedView
+	}
+
+	tk.MustExec("alter materialized view mv comment = 'updated comment'")
+	mvTable, mvInfo := getMView()
+	require.Equal(t, "updated comment", mvTable.Comment)
+
+	tk.MustExec("alter materialized view mv refresh start with date_add(now(), interval 40 minute) next date_add(now(), interval 20 minute)")
+	_, mvInfo = getMView()
+	require.Equal(t, "FAST", mvInfo.RefreshMethod)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 40 MINUTE)", mvInfo.RefreshStartWith)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 20 MINUTE)", mvInfo.RefreshNext)
+	tk.MustQuery(fmt.Sprintf("select NEXT_REFRESH_UNIX_SECONDS is not null from mysql.tidb_mview_refresh_info where MVIEW_ID = %d", mvTable.ID)).Check(testkit.Rows("1"))
+
+	tk.MustExec("alter materialized view mv attributes='mview_alert_warning=5,mview_alert_overdue=10,mview_alert_refresh_failed=yes'")
+	_, mvInfo = getMView()
+	require.Equal(t, int64(5), mvInfo.AlertWarningSec)
+	require.Equal(t, int64(10), mvInfo.AlertOverdueSec)
+	require.True(t, mvInfo.AlertRefreshFailed)
+
+	err := tk.ExecToErr("alter materialized view mv attributes='mview_alert_warning=20,mview_alert_overdue=10'")
+	require.ErrorContains(t, err, "must be less than or equal")
+
+	tk.MustExec("alter materialized view mv refresh")
+	_, mvInfo = getMView()
+	require.Empty(t, mvInfo.RefreshStartWith)
+	require.Empty(t, mvInfo.RefreshNext)
+	tk.MustQuery(fmt.Sprintf("select NEXT_REFRESH_UNIX_SECONDS is null from mysql.tidb_mview_refresh_info where MVIEW_ID = %d", mvTable.ID)).Check(testkit.Rows("1"))
+}
+
+func TestAlterMaterializedViewAttributesUpdatesAlertThresholds(t *testing.T) {
+	store, dom := testkit.CreateMockStoreAndDomain(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t_mv_attr (a int not null, b int not null)")
+	tk.MustExec("create materialized view log on t_mv_attr (a, b)")
+	tk.MustExec("create materialized view mv_attr (a, s, cnt) refresh fast next date_add(now(), interval 2 hour) attributes='mview_alert_warning=300,mview_alert_overdue=600,mview_alert_refresh_failed=no' as select a, sum(b), count(1) from t_mv_attr group by a")
+
+	getMViewInfo := func() *model.MaterializedViewInfo {
+		tbl, err := dom.InfoSchema().TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("mv_attr"))
+		require.NoError(t, err)
+		require.NotNil(t, tbl.Meta().MaterializedView)
+		return tbl.Meta().MaterializedView
+	}
+
+	mvInfo := getMViewInfo()
+	require.Equal(t, int64(300), mvInfo.AlertWarningSec)
+	require.Equal(t, int64(600), mvInfo.AlertOverdueSec)
+	require.False(t, mvInfo.AlertRefreshFailed)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 2 HOUR)", mvInfo.RefreshNext)
+
+	tk.MustExec("alter materialized view mv_attr attributes='mview_alert_warning=5,mview_alert_overdue=5,mview_alert_refresh_failed=yes'")
+	mvInfo = getMViewInfo()
+	require.Equal(t, int64(5), mvInfo.AlertWarningSec)
+	require.Equal(t, int64(5), mvInfo.AlertOverdueSec)
+	require.True(t, mvInfo.AlertRefreshFailed)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 2 HOUR)", mvInfo.RefreshNext)
+
+	tk.MustExec("alter materialized view mv_attr refresh next date_add(now(), interval 25 minute)")
+	tk.MustExec("alter materialized view mv_attr attributes='mview_alert_warning=10,mview_alert_overdue=20,mview_alert_refresh_failed=no'")
+	mvInfo = getMViewInfo()
+	require.Equal(t, int64(10), mvInfo.AlertWarningSec)
+	require.Equal(t, int64(20), mvInfo.AlertOverdueSec)
+	require.False(t, mvInfo.AlertRefreshFailed)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 25 MINUTE)", mvInfo.RefreshNext)
+
+	err := tk.ExecToErr("alter materialized view mv_attr attributes='unknown=1'")
+	require.ErrorContains(t, err, "unsupported ATTRIBUTES key")
+	err = tk.ExecToErr("alter materialized view mv_attr attributes='mview_alert_refresh_failed=maybe'")
+	require.ErrorContains(t, err, "must be yes or no")
+	err = tk.ExecToErr("alter materialized view mv_attr attributes='mview_alert_warning=20,mview_alert_overdue=10'")
+	require.ErrorContains(t, err, "must be less than or equal")
+}
+
+func TestAlterMaterializedViewRefreshDisableScheduleUpdatesAlert(t *testing.T) {
+	tests := []struct {
+		name           string
+		refreshFailed  bool
+		expectAlertRow bool
+	}{
+		{name: "clears non-failed alert"},
+		{name: "preserves refresh-failed alert", refreshFailed: true, expectAlertRow: true},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			store, dom := testkit.CreateMockStoreAndDomain(t)
+			tk := newMViewTestKit(t, store)
+			tk.MustExec("use test")
+			tk.MustExec("create table t_mv_alert (a int not null, b int not null)")
+			tk.MustExec("create materialized view log on t_mv_alert (a, b)")
+			tk.MustExec("create materialized view mv_alert (a, s, cnt) refresh fast next date_add(now(), interval 2 hour) as select a, sum(b), count(1) from t_mv_alert group by a")
+
+			tbl, err := dom.InfoSchema().TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("mv_alert"))
+			require.NoError(t, err)
+			refreshFailed := "NULL"
+			if tt.refreshFailed {
+				refreshFailed = "'YES'"
+			}
+			tk.MustExec(fmt.Sprintf(
+				"insert into mysql.tidb_mview_refresh_alert (MVIEW_ID, MVIEW_SCHEMA, MVIEW_NAME, ALERT_LEVEL, REFRESH_FAILED, LAST_SUCCESS_SNAPSHOT_TIME, UPDATE_TIME) values (%d, 'test', 'mv_alert', 'warning', %s, UTC_TIMESTAMP(), UTC_TIMESTAMP())",
+				tbl.Meta().ID,
+				refreshFailed,
+			))
+
+			tk.MustExec("alter materialized view mv_alert refresh")
+			tk.MustQuery(fmt.Sprintf("select NEXT_REFRESH_UNIX_SECONDS is null from mysql.tidb_mview_refresh_info where MVIEW_ID = %d", tbl.Meta().ID)).Check(testkit.Rows("1"))
+			if !tt.expectAlertRow {
+				tk.MustQuery(fmt.Sprintf("select count(*) from mysql.tidb_mview_refresh_alert where MVIEW_ID = %d", tbl.Meta().ID)).Check(testkit.Rows("0"))
+				return
+			}
+			tk.MustQuery(fmt.Sprintf("select ALERT_LEVEL is null, REFRESH_FAILED from mysql.tidb_mview_refresh_alert where MVIEW_ID = %d", tbl.Meta().ID)).Check(testkit.Rows("1 YES"))
+		})
+	}
+}
+
+func TestAlterMaterializedViewRefreshScheduleTimeZone(t *testing.T) {
+	store, dom := testkit.CreateMockStoreAndDomain(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("set time_zone = '+00:00'")
+	tk.MustExec("create table t (a int not null, b int not null)")
+	tk.MustExec("create materialized view log on t (a, b)")
+	tk.MustExec("create materialized view mv (a, s, cnt) refresh fast next cast('2030-01-01 10:00:00' as datetime) as select a, sum(b), count(1) from t group by a")
+
+	getMView := func() (*model.TableInfo, *model.MaterializedViewInfo) {
+		is := dom.InfoSchema()
+		tbl, err := is.TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("mv"))
+		require.NoError(t, err)
+		require.NotNil(t, tbl.Meta().MaterializedView)
+		return tbl.Meta(), tbl.Meta().MaterializedView
+	}
+
+	mvTable, info := getMView()
+	initialTimeZoneName := info.RefreshScheduleTimeZone.Name
+	initialTimeZoneOffset := info.RefreshScheduleTimeZone.Offset
+	require.Equal(t, 0, initialTimeZoneOffset)
+
+	tk.MustExec("set time_zone = '+08:00'")
+	tk.MustExec("alter materialized view mv refresh")
+	_, info = getMView()
+	require.Equal(t, initialTimeZoneName, info.RefreshScheduleTimeZone.Name)
+	require.Equal(t, initialTimeZoneOffset, info.RefreshScheduleTimeZone.Offset)
+	require.Empty(t, info.RefreshNext)
+
+	tk.MustExec("alter materialized view mv refresh next cast('2030-01-02 10:00:00' as datetime)")
+	_, info = getMView()
+	require.Equal(t, 8*60*60, info.RefreshScheduleTimeZone.Offset)
+	tk.MustQuery("select NEXT_REFRESH_UNIX_SECONDS = 1893549600 from mysql.tidb_mview_refresh_info where MVIEW_ID = " + strconv.FormatInt(mvTable.ID, 10)).Check(testkit.Rows("1"))
+}
+
+func TestAlterMaterializedViewLogDDL(t *testing.T) {
+	store, dom := testkit.CreateMockStoreAndDomain(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t (a int not null, b varchar(10) not null, c int)")
+	tk.MustExec("create materialized view log on t (a) purge next date_add(now(), interval 2 hour)")
+	tk.MustExec("alter materialized view log on t add column (b, c)")
+
+	getMLog := func() (*model.TableInfo, *model.MaterializedViewLogInfo) {
+		is := dom.InfoSchema()
+		tbl, err := is.TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("$mlog$t"))
+		require.NoError(t, err)
+		require.NotNil(t, tbl.Meta().MaterializedViewLog)
+		return tbl.Meta(), tbl.Meta().MaterializedViewLog
+	}
+
+	mlogTable, mlogInfo := getMLog()
+	require.Equal(t, []ast.CIStr{ast.NewCIStr("a"), ast.NewCIStr("b"), ast.NewCIStr("c")}, mlogInfo.Columns)
+	columnNames := make([]string, 0, len(mlogTable.Columns))
+	for _, col := range mlogTable.Columns {
+		columnNames = append(columnNames, col.Name.O)
+	}
+	require.Equal(t, []string{"a", "b", "c", "_MLOG$_DML_TYPE", "_MLOG$_OLD_NEW"}, columnNames)
+
+	err := tk.ExecToErr("alter materialized view log on t add column (b)")
+	require.ErrorContains(t, err, "Duplicate column name")
+	err = tk.ExecToErr("alter materialized view log on t add column (missing_col)")
+	require.ErrorContains(t, err, "Unknown column")
+
+	tk.MustExec("alter materialized view log on t purge start with date_add(now(), interval 40 minute) next date_add(now(), interval 20 minute)")
+	mlogTable, mlogInfo = getMLog()
+	require.Equal(t, "DEFERRED", mlogInfo.PurgeMethod)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 40 MINUTE)", mlogInfo.PurgeStartWith)
+	require.Equal(t, "DATE_ADD(NOW(), INTERVAL 20 MINUTE)", mlogInfo.PurgeNext)
+	tk.MustQuery(fmt.Sprintf("select NEXT_PURGE_UNIX_SECONDS is not null from mysql.tidb_mlog_purge_info where MLOG_ID = %d", mlogTable.ID)).Check(testkit.Rows("1"))
+
+	tk.MustExec("alter materialized view log on t purge")
+	_, mlogInfo = getMLog()
+	require.Empty(t, mlogInfo.PurgeStartWith)
+	require.Empty(t, mlogInfo.PurgeNext)
+	tk.MustQuery(fmt.Sprintf("select NEXT_PURGE_UNIX_SECONDS is null from mysql.tidb_mlog_purge_info where MLOG_ID = %d", mlogTable.ID)).Check(testkit.Rows("1"))
+}
+
+func TestAlterMaterializedViewLogPurgeExprTypeValidation(t *testing.T) {
+	store := testkit.CreateMockStore(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t (a int, b int)")
+	tk.MustExec("create materialized view log on t (a) purge next date_add(now(), interval 1 hour)")
+
+	err := tk.ExecToErr("alter materialized view log on t purge start with 1 next date_add(now(), interval 1 hour)")
+	require.ErrorContains(t, err, "PURGE START WITH expression must return DATETIME/TIMESTAMP")
+
+	err = tk.ExecToErr("alter materialized view log on t purge next 300")
+	require.ErrorContains(t, err, "PURGE NEXT expression must return DATETIME/TIMESTAMP")
+
+	err = tk.ExecToErr("alter materialized view log on t purge immediate")
+	require.ErrorContains(t, err, "PURGE IMMEDIATE is not supported for ALTER MATERIALIZED VIEW LOG")
+
+	tk.MustExec("alter materialized view log on t purge start with now() next date_add(now(), interval 1 hour)")
+}
+
+func TestAlterMaterializedViewLogPurgeScheduleTimeZone(t *testing.T) {
+	store, dom := testkit.CreateMockStoreAndDomain(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("set time_zone = '+00:00'")
+	tk.MustExec("create table t (a int)")
+	tk.MustExec("create materialized view log on t (a) purge next cast('2030-01-01 10:00:00' as datetime)")
+
+	getMLog := func() (*model.TableInfo, *model.MaterializedViewLogInfo) {
+		is := dom.InfoSchema()
+		tbl, err := is.TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("$mlog$t"))
+		require.NoError(t, err)
+		require.NotNil(t, tbl.Meta().MaterializedViewLog)
+		return tbl.Meta(), tbl.Meta().MaterializedViewLog
+	}
+
+	mlogTable, info := getMLog()
+	initialTimeZoneName := info.PurgeScheduleTimeZone.Name
+	initialTimeZoneOffset := info.PurgeScheduleTimeZone.Offset
+	require.Equal(t, 0, initialTimeZoneOffset)
+
+	tk.MustExec("set time_zone = '+08:00'")
+	tk.MustExec("alter materialized view log on t purge")
+	_, info = getMLog()
+	require.Equal(t, initialTimeZoneName, info.PurgeScheduleTimeZone.Name)
+	require.Equal(t, initialTimeZoneOffset, info.PurgeScheduleTimeZone.Offset)
+	require.Empty(t, info.PurgeNext)
+
+	tk.MustExec("alter materialized view log on t purge next cast('2030-01-02 10:00:00' as datetime)")
+	_, info = getMLog()
+	require.Equal(t, 8*60*60, info.PurgeScheduleTimeZone.Offset)
+	tk.MustQuery("select NEXT_PURGE_UNIX_SECONDS = 1893549600 from mysql.tidb_mlog_purge_info where MLOG_ID = " + strconv.FormatInt(mlogTable.ID, 10)).Check(testkit.Rows("1"))
+}
+
+func TestAlterMaterializedViewLogAddColumnRejectsInvalidColumns(t *testing.T) {
+	store := testkit.CreateMockStore(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t_add_mlog_invalid (a int, b int, c int)")
+	tk.MustExec("create materialized view log on t_add_mlog_invalid (a)")
+
+	tk.MustGetErrCode("alter materialized view log on t_add_mlog_invalid add column (a)", errno.ErrDupFieldName)
+	tk.MustGetErrCode("alter materialized view log on t_add_mlog_invalid add column (b, b)", errno.ErrDupFieldName)
+	tk.MustGetErrCode("alter materialized view log on t_add_mlog_invalid add column (missing_col)", errno.ErrBadField)
+	tk.MustGetErrCode("alter materialized view log on t_add_mlog_invalid add column (`_MLOG$_DML_TYPE`)", errno.ErrDupFieldName)
+	tk.MustGetErrMsg(
+		"alter materialized view log on t_add_mlog_invalid add column (b), add column (c)",
+		"[ddl:8200]Unsupported ALTER MATERIALIZED VIEW LOG with multiple ADD COLUMN actions",
+	)
+
+	tk.MustExec("create table t_add_mlog_unsupported (id int, b blob, j json, g1 int, g2 int as (g1 + 1) stored)")
+	tk.MustExec("create materialized view log on t_add_mlog_unsupported (id)")
+	err := tk.ExecToErr("alter materialized view log on t_add_mlog_unsupported add column (b)")
+	require.ErrorContains(t, err, "ALTER MATERIALIZED VIEW LOG does not support BLOB column b")
+	err = tk.ExecToErr("alter materialized view log on t_add_mlog_unsupported add column (j)")
+	require.ErrorContains(t, err, "ALTER MATERIALIZED VIEW LOG does not support JSON column j")
+	tk.MustExec("alter materialized view log on t_add_mlog_unsupported add column (g2)")
+}
+
+func TestAlterMaterializedViewLogAddColumnCopiesBaseColumnMetadata(t *testing.T) {
+	store, dom := testkit.CreateMockStoreAndDomain(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t_add_mlog_col (id int not null, n int not null, s varchar(10) not null, d date not null, note text not null)")
+	tk.MustExec("create materialized view log on t_add_mlog_col (id)")
+	tk.MustExec("alter materialized view log on t_add_mlog_col add column (n, s, d, note)")
+
+	is := dom.InfoSchema()
+	mlogTable, err := is.TableByName(context.Background(), ast.NewCIStr("test"), ast.NewCIStr("$mlog$t_add_mlog_col"))
+	require.NoError(t, err)
+	require.Equal(t, []ast.CIStr{ast.NewCIStr("id"), ast.NewCIStr("n"), ast.NewCIStr("s"), ast.NewCIStr("d"), ast.NewCIStr("note")}, mlogTable.Meta().MaterializedViewLog.Columns)
+
+	columns := make(map[string]*model.ColumnInfo, len(mlogTable.Meta().Columns))
+	for _, column := range mlogTable.Meta().Columns {
+		columns[column.Name.L] = column
+	}
+	for _, name := range []string{"n", "s", "d", "note"} {
+		require.True(t, mysql.HasNotNullFlag(columns[name].GetFlag()))
+	}
+	require.Equal(t, mysql.TypeBlob, columns["note"].GetType())
+}
+
+func TestAlterMaterializedViewLogAddColumnSupportsNewMaterializedView(t *testing.T) {
+	store := testkit.CreateMockStore(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t_add_mlog_mv (a int not null, b int not null, c int not null)")
+	tk.MustExec("insert into t_add_mlog_mv values (1, 10, 100), (1, 20, 200), (2, 30, 300)")
+	tk.MustExec("create materialized view log on t_add_mlog_mv (a, b)")
+
+	err := tk.ExecToErr("create materialized view mv_add_mlog_col_before (a, s, cnt) refresh fast as select a, sum(c), count(1) from t_add_mlog_mv group by a")
+	require.ErrorContains(t, err, "materialized view log does not contain column c")
+
+	tk.MustExec("alter materialized view log on t_add_mlog_mv add column (c)")
+	tk.MustExec("create materialized view mv_add_mlog_col_after (a, s, cnt) refresh fast as select a, sum(c), count(1) from t_add_mlog_mv group by a")
+	tk.MustQuery("select a, s, cnt from mv_add_mlog_col_after order by a").Check(testkit.Rows("1 300 2", "2 300 1"))
+}
+
+func TestAlterTableWithMaterializedViewDependencies(t *testing.T) {
+	store := testkit.CreateMockStore(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t (a int not null, b int not null)")
+	tk.MustExec("insert into t values (1, 10), (2, -1)")
+	tk.MustExec("create materialized view log on t (a, b)")
+	tk.MustExec("create materialized view mv (a, cnt) as select a, count(1) from t group by a")
+	tk.MustExec("create materialized view mv_where (a, cnt) as select a, count(1) from t where b > 0 group by a")
+
+	tk.MustExec("alter table t modify column a bigint not null")
+	showCreate := tk.MustQuery("show create table `$mlog$t`").Rows()[0][1].(string)
+	require.Contains(t, showCreate, "`a` bigint")
+	showCreate = tk.MustQuery("show create table mv").Rows()[0][1].(string)
+	require.Contains(t, showCreate, "`a` bigint")
+
+	err := tk.ExecToErr("alter table t modify column b bigint not null")
+	require.ErrorContains(t, err, "does not support modifying columns used in WHERE clause")
+	err = tk.ExecToErr("alter table t change column a a2 bigint")
+	require.ErrorContains(t, err, "does not support renaming")
+	err = tk.ExecToErr("alter table t modify column b smallint")
+	require.ErrorContains(t, err, "only supports no-reorg compatible type changes")
+}
+
+func TestMaterializedViewRelatedTablesDDLRejected(t *testing.T) {
+	store := testkit.CreateMockStore(t)
+	tk := newMViewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t_ddl_mv (a int not null, b int)")
+	tk.MustExec("create materialized view log on t_ddl_mv (a, b)")
+
+	err := tk.ExecToErr("drop table t_ddl_mv")
+	require.ErrorContains(t, err, "DROP TABLE on base table with materialized view log")
+	err = tk.ExecToErr("rename table t_ddl_mv to t_ddl_mv2")
+	require.ErrorContains(t, err, "RENAME TABLE on base table with materialized view log")
+	err = tk.ExecToErr("drop table `$mlog$t_ddl_mv`")
+	require.ErrorContains(t, err, "DROP TABLE on materialized view log table")
+
+	tk.MustExec("create materialized view mv_ddl_mv (a, cnt) as select a, count(1) from t_ddl_mv group by a")
+	tk.MustExec("alter table t_ddl_mv add column c int")
+	err = tk.ExecToErr("alter table t_ddl_mv modify column a bigint")
+	require.ErrorContains(t, err, "does not support changing charset/collation/nullability of group keys")
+	err = tk.ExecToErr("drop table t_ddl_mv")
+	require.ErrorContains(t, err, "DROP TABLE on base table with materialized view dependencies")
+	err = tk.ExecToErr("alter table mv_ddl_mv add column x int")
+	require.ErrorContains(t, err, "ALTER TABLE on materialized view table")
+	err = tk.ExecToErr("alter table `$mlog$t_ddl_mv` add index idx_mlog_b(b)")
+	require.ErrorContains(t, err, "ALTER TABLE on materialized view log table")
+	err = tk.ExecToErr("create index idx_mlog_b_create on `$mlog$t_ddl_mv`(b)")
+	require.ErrorContains(t, err, "CREATE INDEX on materialized view log table")
 }
